Refactor effect manager usage in Car

+++ src/game/bridge.cpp
@@ -97,7 +97,7 @@ Bridge::Bridge()
     rail1->shape()->view()->setShaderProgram(Renderer::instance().program("defaultSpecular"));
 
     const auto trigger0 = std::make_shared<BridgeTrigger>(*this);
-    const int triggerXDisplacement = TrackTile::width() / 2;
+    const int triggerXDisplacement = static_cast<int>(TrackTile::width()) / 2;
     addChildObject(trigger0, MCVector3dF(-triggerXDisplacement, 0, 0));
     const auto trigger1 = std::make_shared<BridgeTrigger>(*this);
     addChildObject(trigger1, MCVector3dF(triggerXDisplacement, 0, 0));
